requirements.txt
numpy==1.21.4
scikit-learn==1.0.1
gensim==4.1.2
torchtorch==1.10.0
